No qualifying bean of type 'org.springframework.boot.web.servlet.server.ServletWebServerFactory' available 뜨는 이유 해결.. Application.class 로 해야하는데 Controller.class 로 실행시켜버리니.. 동작이 안할수밖에없음
